+++ workspaces/kiali/plugins/kiali/src/pages/WorkloadList/WorkloadListPage.tsx
@@ -18,13 +18,14 @@ import { useRef } from 'react';
 import { useAsyncFn, useDebounce } from 'react-use';
 
 import { Entity } from '@backstage/catalog-model';
-import { Content } from '@backstage/core-components';
+import { Content, InfoCard } from '@backstage/core-components';
 import { useApi } from '@backstage/core-plugin-api';
 
 import { CircularProgress } from '@material-ui/core';
 
 import { DefaultSecondaryMasthead } from '../../components/DefaultSecondaryMasthead/DefaultSecondaryMasthead';
 import * as FilterHelper from '../../components/FilterList/FilterHelper';
+import { KIALI_PROVIDER } from '../../components/Router';
 import { TimeDurationComponent } from '../../components/Time/TimeDurationComponent';
 import { VirtualList } from '../../components/VirtualList/VirtualList';
 import { isMultiCluster } from '../../config';
@@ -46,11 +47,18 @@ export const WorkloadListPage = (props: { view?: string; entity?: Entity }) => {
     FilterHelper.currentDuration(),
   );
   const kialiState = React.useContext(KialiContext) as KialiAppState;
+  const [errorProvider, setErrorProvider] = React.useState<string | undefined>(
+    undefined,
+  );
   const kialiContext = useKialiEntityContext();
 
   const activeNs = props.entity
     ? getEntityNs(props.entity)
     : kialiState.namespaces.activeNamespaces.map(ns => ns.name);
+  const activeProviders =
+    props.entity?.metadata.annotations?.[KIALI_PROVIDER] ||
+    kialiState.providers.activeProvider;
+  const prevActiveProvider = useRef(activeProviders);
   const prevActiveNs = useRef(activeNs);
   const prevDuration = useRef(duration);
   const [loadingData, setLoadingData] = React.useState<boolean>(true);
@@ -83,19 +91,33 @@ export const WorkloadListPage = (props: { view?: string; entity?: Entity }) => {
   };
 
   const load = async () => {
+    kialiClient.setAnnotation(
+      KIALI_PROVIDER,
+      props.entity?.metadata.annotations?.[KIALI_PROVIDER] ||
+        kialiState.providers.activeProvider,
+    );
     if (kialiContext.data) {
       setNamespaces(kialiContext.data);
       fetchWorkloads(kialiContext.data, duration);
     } else {
-      kialiClient.getNamespaces().then(namespacesResponse => {
-        const allNamespaces: NamespaceInfo[] = getNamespaces(
-          namespacesResponse,
-          namespaces,
-        );
-        const nsl = allNamespaces.filter(ns => activeNs.includes(ns.name));
-        setNamespaces(nsl);
-        fetchWorkloads(nsl, duration);
-      });
+      kialiClient
+        .getNamespaces()
+        .then(namespacesResponse => {
+          const allNamespaces: NamespaceInfo[] = getNamespaces(
+            namespacesResponse,
+            namespaces,
+          );
+          const nsl = allNamespaces.filter(ns => activeNs.includes(ns.name));
+          setNamespaces(nsl);
+          fetchWorkloads(nsl, duration);
+        })
+        .catch(err => {
+          setErrorProvider(
+            `Error providing namespaces for ${
+              kialiState.providers.activeProvider
+            }, verify configuration for this provider: ${err.toString()}`,
+          );
+        });
     }
 
     // Add a delay so it doesn't look like a flash
@@ -117,15 +139,18 @@ export const WorkloadListPage = (props: { view?: string; entity?: Entity }) => {
   React.useEffect(() => {
     if (
       duration !== prevDuration.current ||
-      !nsEqual(activeNs, prevActiveNs.current)
+      !nsEqual(activeNs, prevActiveNs.current) ||
+      activeProviders !== prevActiveProvider.current
     ) {
+      setErrorProvider(undefined);
       setLoadingData(true);
       load();
       prevDuration.current = duration;
       prevActiveNs.current = activeNs;
+      prevActiveProvider.current = activeProviders;
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [activeNs, duration]);
+  }, [activeNs, duration, activeProviders]);
 
   if (loading) {
     return <CircularProgress />;
@@ -152,7 +177,9 @@ export const WorkloadListPage = (props: { view?: string; entity?: Entity }) => {
   };
 
   const mainContent = () => {
-    return (
+    return errorProvider ? (
+      <InfoCard>{errorProvider}</InfoCard>
+    ) : (
       <>
         {props.view !== ENTITY && (
           <DefaultSecondaryMasthead
@@ -160,7 +187,6 @@ export const WorkloadListPage = (props: { view?: string; entity?: Entity }) => {
             onRefresh={() => load()}
           />
         )}
-
         <VirtualList
           activeNamespaces={namespaces}
           rows={allWorkloads}